local user additions
code clean up
added functions for local user management

+++ finalproject.py
@@ -1,13 +1,29 @@
 from flask import Flask, render_template,\
     request, redirect, url_for, flash, jsonify
+from sqlalchemy import create_engine, and_, desc
+from sqlalchemy.orm import sessionmaker
+from database_setup import BASE, Restaurant, MenuItem, User
+
+from flask import session as login_session
+import random
+import string
+
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+from oauth2client.client import AccessTokenCredentials
+
+import httplib2
+import json
+from flask import make_response
+import requests
+
 
 app = Flask(__name__)
 
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Restaurant Menu Application"
 
-
-from sqlalchemy import create_engine, and_, desc
-from sqlalchemy.orm import sessionmaker
-from database_setup import BASE, Restaurant, MenuItem
 
 engine = create_engine('sqlite:///restaurantmenu.db')
 BASE.metadata.bind = engine
@@ -15,10 +31,178 @@ BASE.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-#restaurants = [{'name':'franks','id':0},{'name':'mark','id':1}]
-#menu = [{}]
 
-#Show all Restaurants
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session['email'],
+                   picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_credentials = login_session.get('credentials')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_credentials is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['credentials'] = credentials
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+    data = answer.json()
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+    login_session['access_token'] = credentials.access_token
+    print login_session, "mark 1"
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += """ " style = "width: 300px; height: 300px;border-radius: 
+    150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> """
+    flash("you are now logged in as %s" % login_session['username'])
+    user = getUserID(login_session['email'])
+    if user is None:
+        user_id = createUser(login_session)
+        login_session['user_id'] = user_id
+    else:
+        login_session['user_id'] = getUserInfo(User.id)
+    print "done!"
+    return output
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session.get('access_token')
+    print 'In gdisconnect access token is %s' % str(access_token)
+    print 'User name is: '
+    print login_session.get('username')
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps(
+            'Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % str(
+        login_session.get('access_token'))
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result, login_session, "this is"
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps(
+            'Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+# JSON APIs to view Restaurant Information
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/JSON')
+def restaurantMenuJSON(restaurant_id):
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    items = session.query(MenuItem).filter_by(
+        restaurant_id=restaurant_id).all()
+    return jsonify(MenuItems=[i.serialize for i in items])
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/JSON')
+def menuItemJSON(restaurant_id, menu_id):
+    Menu_Item = session.query(MenuItem).filter_by(id=menu_id).one()
+    return jsonify(Menu_Item=Menu_Item.serialize)
+
+
+# Show all Restaurants
 @app.route('/')
 @app.route('/restaurant')
 def showRestaurants():
@@ -27,10 +211,12 @@ def showRestaurants():
     return render_template('index.html', restaurants=restaurants)
 
 
-#Create new restrant 
+# Create new restrant
 @app.route('/restaurant/new/', methods=['GET', 'POST'])
 def newRestaurant():
     """ show new Restaurant form """
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     if request.method == 'POST':
         req = request.form
         restaurant = Restaurant(name=req['name'])
@@ -41,10 +227,14 @@ def newRestaurant():
     else:
         return render_template('newrestaurant.html')
 
-#edit Restaurant
+# edit Restaurant
+
+
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods=['GET', 'POST'])
 def editRestaurant(restaurant_id):
     """ edit Restaurant"""
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     restaurant = session.query(Restaurant).get(restaurant_id)
     if request.method == 'POST':
         req = request.form
@@ -55,10 +245,14 @@ def editRestaurant(restaurant_id):
     else:
         return render_template('editrestaurant.html', restaurant=restaurant)
 
-#delete Restaurant
+# delete Restaurant
+
+
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods=['GET', 'POST'])
 def deleteRestaurant(restaurant_id):
     """ Delete Restaurant """
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     restaurant = session.query(Restaurant).get(restaurant_id)
     if request.method == 'POST':
         session.delete(restaurant)
@@ -69,19 +263,26 @@ def deleteRestaurant(restaurant_id):
         return render_template('deleterestaurant.html', restaurant=restaurant)
 
 # Show a restaurant menu
+
+
 @app.route('/restaurant/<int:restaurant_id>/')
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     """ show menu """
     restaurant = session.query(Restaurant).get(restaurant_id)
-    items = session.query(MenuItem).filter_by(restaurant_id=restaurant_id).all()
+    items = session.query(MenuItem).filter_by(
+        restaurant_id=restaurant_id).all()
     return render_template('showmenu.html', items=items, restaurant=restaurant)
 
 # Create a new menu item
+
+
 @app.route(
     '/restaurant/<int:restaurant_id>/menu/new/', methods=['GET', 'POST'])
 def newMenuItem(restaurant_id):
     """ New Menu Item"""
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     if request.method == 'POST':
         req = request.form
         item = MenuItem(name=req['name'], course=req['course'],
@@ -95,10 +296,14 @@ def newMenuItem(restaurant_id):
         return render_template('newmenuitem.html', restaurant_id=restaurant_id)
 
 # Edit a menu item
+
+
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit',
            methods=['GET', 'POST'])
 def editMenuItem(restaurant_id, menu_id):
     """ Edit Menu Item """
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     item = session.query(MenuItem).filter(
         MenuItem.restaurant_id == restaurant_id).filter(MenuItem.id == menu_id).one()
     if request.method == 'POST':
@@ -113,10 +318,14 @@ def editMenuItem(restaurant_id, menu_id):
         return render_template('editmenuitem.html', item=item)
 
 # Delete a menu item
+
+
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete',
            methods=['GET', 'POST'])
 def deleteMenuItem(restaurant_id, menu_id):
     """ Delete Menu Item """
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     if request.method == 'POST':
         item = session.query(MenuItem).filter(
             MenuItem.restaurant_id == restaurant_id).filter(MenuItem.id == menu_id).one()
